Move config routes into standalone file
