+++ views/config.ejs
@@ -128,7 +128,7 @@
    <!-- END TABLE-->
 
    <!-- Modal Add Configuration-->
-	  <form action="/config/save" method="post">
+	  <form action="/app/config/save" method="post">
         <div class="modal fade" id="myModalAdd" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
          <div class="modal-dialog" role="document">
            <div class="modal-content">
@@ -165,7 +165,7 @@
    <!-- End Modal Add Configuration-->
 
    <!-- Modal Update Configuration-->
-   <form action="/config/update" method="post">
+   <form action="/app/config/update" method="post">
        <div class="modal fade" id="EditModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
         <div class="modal-dialog" role="document">
           <div class="modal-content">
@@ -203,7 +203,7 @@
    <!-- End Modal Update Configuration-->
 
 	 <!-- Modal Delete Configuration-->
- 	  <form id="add-row-form" action="/config/delete" method="post">
+ 	  <form id="add-row-form" action="/app/config/delete" method="post">
  	     <div class="modal fade" id="DeleteModal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
  	        <div class="modal-dialog">
  	           <div class="modal-content">
